Check each transit day for whether it's a delivery day rather than just the final day.
Previously we only checked the delivery day against the "delivery days" setting.
This meant that a 5 day estimate that includes a weekend but ends on a weekday provides
an earlier delivery estimate than you would expect. You wouldn't expect to include the weekend
when counting down the 5 days.  Instead of adding a new "transitable" days estimate I have
checked as we count each day of the estimate against whether it's a delivery day and considered
this synonymous to whether the courier is travelling that day.